Eleventh push 2/2. CharacterSheet: some of the logic behind adding nodes to the TreeView and the character's family has been moved to an external service, creating a new View for the purpose. - Still have to clean the Sheet from lots of logic.
With this migration I also changed the behavior for the TreeView drawing, that was static and hardcoded; it now allows the addition of new ties, but the calculations behind made by the Syncer are still hardcoded.
Yet, for adding a new tie now all that's needed is to add it to the Variables class and edit the Syncer.