Thirteenth push 4/2.
Eliminated the enums, replacing them by a service that loads and stores variables from external files, opening the app to future modifications for Genders, Races, Conditions, Special Conditions, and Ties. - Corrected the Syncer to work with this new service, eliminating the harding and making it flexible.

Added the Mapper app to the Maro app. Started the application of MVP to the Mapper.